Create QUESTION FOR GRADER IN DESCRIPTION
The instructions weren't super clear about what the final product was supposed to be, or maybe it's my bad English. From the quiz, it sounded like the simple changes I made for a11y, SEO, and semantics 
were all the were required. If we were supposed to do more, please let me know.